Backend/requirements.txt
# FastAPI Backend for ChatTax
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
bcrypt==4.2.1
email-validator==2.2.0

# SSE & Async (compatible version with FastAPI 0.115.0)
sse-starlette==1.8.2
httpx==0.28.1

# LangChain & LLM
langchain==0.3.13
langchain-openai==0.2.14
langchain-community==0.3.13
openai==1.58.1

# Vector Database & Embeddings
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1

# Reranking
# Using sentence-transformers for cross-encoder models

# Data Processing
pandas==2.2.3
pyarrow==17.0.0

# Utils
python-dateutil==2.9.0
tiktoken==0.8.0
